Fix: UI Consistency in InfoCard
